requirements.txt
absl-py==2.2.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
certifi==2023.7.22
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Django==5.2.2
django-cors-headers==4.9.0
django-crontab==0.7.1
django-extensions==4.1
django-filter==25.2
django-prometheus==2.4.1
django-rest-framework==0.1.0
django-reversion==6.0.0
djangorestframework==3.16.0
djangorestframework-csv==3.0.2
drf-yasg==1.21.11
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.21
grpcio==1.71.0
h5py==3.13.0
httplib2==0.31.0
idna==3.10
inflection==0.5.1
influxdb-client==1.49.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mysqlclient==2.2.7
namex==0.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plantuml==0.3.0
platformdirs==4.3.7
plotly==6.0.1
prometheus_client==0.23.1
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydotplus==2.0.2
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.3
pyzmq==26.3.0
reactivex==4.1.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.5.0
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
